coordinateur_service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
drf-yasg==1.21.7
requests==2.31.0
django-filter==24.1
uuid6